Added nvim version check

+++ core/tabs/applications-setup/Developer-tools/neovim-setup.sh
@@ -5,6 +5,13 @@
 gitpath="$HOME/.local/share/neovim"
 
 cloneNeovim() {
+    # lazy.nvim requires nvim >= 0.8.0
+    nvim_version=$(nvim --version | head -n 1 | awk '{print $2}')
+    if [ "$(printf "%s\n" "$nvim_version" "0.8.0" | sort -V | head -n 1)" != "0.8.0" ]; then
+        printf "%b\n" "${RED}Neovim version $nvim_version not supported.${RC}"
+        exit 1
+    fi
+
     # Check if the dir exists before attempting to clone into it.
     if [ -d "$gitpath" ]; then
         rm -rf "$gitpath"
